added float for k

+++ algorithms/brezenheim.py
@@ -3,7 +3,7 @@ from helper_funcs import calculate_line_points
 from structs.point import Point
 
 
-def brezenheim(color: tuple = RED, k: int = 1, x: int = 30, b: int = 0) -> list[Point]:
+def brezenheim(color: tuple = RED, k: float = 1, x: int = 30, b: int = 0) -> list[Point]:
     '''алгоритм Брезенхема - входные параметры задают уравнение прямой y=kx+b'''
     
     p1, p2 = calculate_line_points(color, k, x, b)
